Move extract status and item types into pkg/extract.
Hooks and folders need a leaf they can import without pulling the daemon.
