+++ pkg/extract/progress.go
@@ -1,0 +1,41 @@
+package extract
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"code.cloudfoundry.org/bytefmt"
+	"golift.io/xtractr"
+)
+
+// Progress holds the progress for an entire Extract.
+// An Extract is "a new item in a watch folder" or "a download in a starr app".
+// Either may produce multiple xtractr.XFile structs (extractable archives).
+type Progress struct {
+	*xtractr.Progress
+	// Extract that exists in the map.
+	*Extract
+	// Number of archives in this Extract.
+	Archives int
+	// Number of archives extracted from this Extract.
+	Extracted int
+}
+
+func (p *Progress) String() string {
+	if p == nil || p.Progress == nil {
+		return "no progress yet"
+	}
+
+	var wrote, total uint64
+
+	if p.Total > 0 {
+		wrote, total = p.Wrote, p.Total
+	} else if p.Compressed > 0 {
+		wrote, total = p.Read, p.Compressed
+	}
+
+	return fmt.Sprintf("on archive: %d/%d @ %sB/%sB (%.0f%%): %s",
+		p.Extracted+1, p.Archives, bytefmt.ByteSize(wrote), bytefmt.ByteSize(total),
+		p.Percent(), strings.TrimLeft(strings.TrimPrefix(p.XFile.FilePath, p.Path), string(filepath.Separator)))
+}
